Split index.js into two files: index.js and css-analyzer.js + add type hinting
